rename bdsqr_host.hpp to rocauxiliary_bdsqr_host.hpp
