+++ library/src/auxiliary/rocauxiliary_bdsqr_host.hpp
